pyproject.toml
[project]
name = "speech-with-slide"
version = "0.1.0"
description = "AI-powered presentation video generator from PDF slides"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PyPDF2>=3.0.1",
    "pdf2image>=1.16.3",
    "pillow>=10.0.0",
    "anthropic>=0.25.0",
    "google-generativeai>=0.7.0",
    "opencv-python>=4.8.0",
    "moviepy>=1.0.3",
    "python-dotenv>=1.0.0",
    "google-genai>=0.8.0",
    "pyyaml>=6.0",
]
